+++ DRN.Framework.EntityFramework/Context/DrnContextServiceRegistrationAttribute.cs
@@ -1,3 +1,4 @@
+using System.Reflection;
 using DRN.Framework.SharedKernel.Enums;
 using DRN.Framework.Utils.DependencyInjection.Attributes;
 using DRN.Framework.Utils.Settings;
@@ -16,11 +17,10 @@ namespace DRN.Framework.EntityFramework.Context;
 /// <see cref="DRN.Framework.Utils.DependencyInjection.ServiceCollectionExtensions.AddServicesWithAttributes"/>
 /// <br/> is called from DbContext's assembly
 /// </summary>
-public class HasDrnContextServiceCollectionModuleAttribute : HasServiceCollectionModuleAttribute
+public class DrnContextServiceRegistrationAttribute : ServiceRegistrationAttribute
 {
-    static HasDrnContextServiceCollectionModuleAttribute() =>
-        ModuleMethodInfo = typeof(ServiceCollectionExtensions)
-            .GetMethod(nameof(ServiceCollectionExtensions.AddDbContextsWithConventions))!;
+    public override void ServiceRegistration(IServiceCollection sc, Assembly? assembly)
+        => sc.AddDbContextsWithConventions(assembly);
 
     public override async Task PostStartupValidationAsync(object service, IServiceProvider serviceProvider)
     {