+++ DRN.Framework.Utils/RELEASE-NOTES.md
@@ -1,5 +1,14 @@
 Not every version includes changes, features or bug fixes. This project can increment version to keep consistency with other DRN.Framework projects.  
 
+## Version 0.3.1
+
+### Breaking Changes
+
+* AttributeSpecifiedServiceCollectionModule renamed as AttributeSpecifiedServiceModule
+* HasServiceCollectionModuleAttribute renamed as ServiceRegistrationAttribute
+* HasDrnContextServiceCollectionModuleAttribute renamed as DrnContextServiceRegistrationAttribute
+* ServiceRegistrationAttribute MethodInfo property replaced with ServiceRegistration method to make usage strongly typed and support inheritance
+
 ## Version 0.3.0
 
 My family celebrates the enduring legacy of Mustafa Kemal Atatürk's enlightenment ideals. This release is dedicated to 23 April National Sovereignty and Children's Day.